Refactor the codebase a bit prior v0.5.0 release
