+++ .github/workflows/tests.yaml
@@ -1,0 +1,30 @@
+name: tests
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+
+jobs:
+  tests:
+    runs-on: ubuntu-latest
+    concurrency:
+      group: ${{ github.workflow }}-${{ github.ref }}
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Install the latest version of uv
+        uses: astral-sh/setup-uv@v5
+        with:
+          version: "latest"
+
+      - name: Set up Python with uv
+        run: uv venv --python 3.11
+
+      - name: Install package and test dependencies
+        run: uv pip install -e ".[test]" --group test
+
+      - name: Run tests
+        run: uv run pytest --tb=short -v
