+++ src/hf_mem/gguf/print.py
@@ -1,123 +1,26 @@
 import warnings
-from functools import cache
-from typing import Any, Dict, Literal, Optional
+from typing import Any, Dict, Optional
 
-from hf_mem import __version__
-from hf_mem.gguf import GGUFMetadata
-from hf_mem.metadata import SafetensorsMetadata
-
-MIN_NAME_LEN = 5
-MAX_NAME_LEN = 14
-MIN_DATA_LEN = 20
-MAX_DATA_LEN = 64
-BORDERS_AND_PADDING = 4
-
-BOX = {
-    "tl": "┌",
-    "tr": "┐",
-    "bl": "└",
-    "br": "┘",
-    "ht": "─",
-    "vt": "│",
-    "tsep": "┬",
-    "bsep": "┴",
-    "lm": "├",
-    "rm": "┤",
-    "mm": "┼",
-}
+from hf_mem._print import (
+    BORDERS_AND_PADDING,
+    MAX_DATA_LEN,
+    MAX_NAME_LEN,
+    _bytes_to_gib,
+    _format_short_number,
+    _make_bar,
+    _print_centered,
+    _print_divider,
+    _print_header,
+    _print_row,
+)
+from hf_mem._version import __version__
+from hf_mem.gguf.metadata import GGUFMetadata
 
 
-def _print_with_color(content: str) -> None:
-    print(f"\x1b[38;2;244;183;63m{content}\x1b[0m")
-
-
-def _print_header(current_len: int) -> None:
-    length = current_len + 2 * BORDERS_AND_PADDING + 2
-    top = BOX["tl"] + (BOX["tsep"] * (length - 2)) + BOX["tr"]
-    _print_with_color(top)
-
-    bottom = BOX["lm"] + (BOX["bsep"] * (length - 2)) + BOX["rm"]
-    _print_with_color(bottom)
-
-
-def _print_centered(text: str, current_len: int) -> None:
-    total_width = current_len + 2 * BORDERS_AND_PADDING
-    text_len = len(text)
-    pad_left = (total_width - text_len) // 2
-    pad_right = total_width - text_len - pad_left
-    _print_with_color(f"{BOX['vt']}{' ' * pad_left}{text}{' ' * pad_right}{BOX['vt']}")
-
-
-def _print_divider(
-    current_len: int,
-    side: Optional[Literal["top", "top-continue", "mid", "bottom", "bottom-continue"]] = None,
-    name_len: int = MAX_NAME_LEN,
-) -> None:
-    match side:
-        case "top":
-            left, mid, right = BOX["lm"], BOX["tsep"], BOX["rm"]
-        case "top-continue":
-            left, mid, right = BOX["lm"], BOX["bsep"], BOX["rm"]
-        case "bottom":
-            left, mid, right = BOX["bl"], BOX["bsep"], BOX["br"]
-        case "bottom-continue":
-            left, mid, right = BOX["lm"], BOX["bsep"], BOX["rm"]
-        case "mid":
-            left, mid, right = BOX["lm"], BOX["ht"], BOX["rm"]
-        case _:
-            left, mid, right = BOX["lm"], BOX["mm"], BOX["rm"]
-
-    name_col_inner = name_len + 2
-    data_col_inner = current_len + 1
-
-    line = left
-    line += BOX["ht"] * name_col_inner
-    line += mid
-    line += BOX["ht"] * data_col_inner
-    line += right
-    _print_with_color(line)
-
-
-def _format_name(name: str, max_len: int = MAX_NAME_LEN) -> str:
-    if len(name) < MIN_NAME_LEN:
-        return f"{name:<{MIN_NAME_LEN}}"
-    if len(name) > max_len:
-        return name[: max_len - 3] + "..."
-    return f"{name:<{max_len}}"
-
-
-def _print_row(name: str, text: str, current_len: int, name_len: int = MAX_NAME_LEN) -> None:
-    name_fmt = _format_name(name, name_len)
-    data_fmt = f"{str(text):<{current_len}}"
-    _print_with_color(f"{BOX['vt']} {name_fmt} {BOX['vt']} {data_fmt} {BOX['vt']}")
-
-
-def _make_bar(used: float, total: float, width: int) -> str:
-    if total <= 0:
-        return "░" * width
-    frac = min(max(used / total, 0.0), 1.0)
-    filled = int(round(frac * width))
-    filled = max(0, min(width, filled))
-    return "█" * filled + "░" * (width - filled)
-
-
-def _format_short_number(n: float) -> str:
-    n = float(n)
-    for unit in ("", "K", "M", "B", "T"):
-        if abs(n) < 1000.0:
-            return f"{int(n)}" if unit == "" else f"{n:.2f}{unit}"
-        n /= 1000.0
-    return f"{n:.2f}P"
-
-
-def _bytes_to_gib(nbytes: int) -> float:
-    return nbytes / (1024**3)
-
-
-def print_report(
+def print_gguf_report(
     model_id: str,
     revision: str,
-    metadata: SafetensorsMetadata | GGUFMetadata,
+    metadata: GGUFMetadata,
     cache: Optional[Dict[str, Any]] = None,
     ignore_table_width: bool = False,
 ) -> None:
@@ -153,7 +56,7 @@ def print_report(
             f"{_bytes_to_gib(metadata.bytes_count):.2f} GiB ({_format_short_number(metadata.param_count)} PARAMS)"
         )
     for _, nested_metadata in metadata.components.items():
-        for dtype, dtype_metadata in nested_metadata.dtypes.items():
+        for _, dtype_metadata in nested_metadata.dtypes.items():
             data_rows.append(
                 f"{_bytes_to_gib(dtype_metadata.bytes_count):.2f} / {_bytes_to_gib(combined_total):.2f} GiB"
             )
@@ -198,6 +101,7 @@ def print_report(
         _print_row("TOTAL MEMORY", model_text, data_col_width)
         _print_row("REQUIREMENTS", model_bar, data_col_width)
 
+    max_length = 0
     for key, value in metadata.components.items():
         if len(metadata.components) > 1:
             _print_divider(data_col_width + 1, "top-continue")
@@ -217,27 +121,19 @@ def print_report(
             _print_divider(data_col_width + 1)
 
         max_length = max(
-            [
-                len(f"{_format_short_number(dtype_metadata.param_count)} PARAMS")
-                for _, dtype_metadata in value.dtypes.items()
-            ]
+            len(f"{_format_short_number(dtype_metadata.param_count)} PARAMS")
+            for _, dtype_metadata in value.dtypes.items()
         )
         for idx, (dtype, dtype_metadata) in enumerate(value.dtypes.items()):
             gib_text = (
                 f"{_bytes_to_gib(dtype_metadata.bytes_count):.2f} / {_bytes_to_gib(combined_total):.2f} GiB"
             )
-            if isinstance(metadata, GGUFMetadata):
-                _print_row(
-                    dtype.name.upper() + " " * (max_length - len(dtype.name)),
-                    gib_text,
-                    data_col_width,
-                )
-            else:
-                _print_row(
-                    dtype.upper() + " " * (max_length - len(dtype)),
-                    gib_text,
-                    data_col_width,
-                )
+            # NOTE: dtype is a GGUFDtype enum — use .name to get the string label
+            _print_row(
+                dtype.name.upper() + " " * (max_length - len(dtype.name)),
+                gib_text,
+                data_col_width,
+            )
 
             bar = _make_bar(
                 _bytes_to_gib(dtype_metadata.bytes_count),
@@ -278,7 +174,7 @@ def print_report(
     _print_divider(data_col_width + 1, "bottom")
 
 
-def print_report_for_gguf(
+def print_gguf_files_report(
     model_id: str,
     revision: str,
     gguf_files: Dict[str, GGUFMetadata],
@@ -297,7 +193,6 @@ def print_report_for_gguf(
     for filename, gguf_metadata in gguf_files.items():
         centered_rows.append(filename)
 
-        # Adding model names lines to compute current_len
         transformer = gguf_metadata.components.get("Transformer")
         if transformer:
             for _, dtype_metadata in transformer.dtypes.items():
@@ -311,7 +206,6 @@ def print_report_for_gguf(
                         f"{filename} | {_bytes_to_gib(dtype_metadata.bytes_count):.2f} GiB ({_format_short_number(dtype_metadata.param_count)} PARAMS)"
                     )
 
-        # Adding also the KV lines to compute current_len
         if gguf_metadata.kv_cache_info is not None:
             cache_size = _bytes_to_gib(gguf_metadata.kv_cache_info.cache_size)
             name_text = f"KV CACHE ({gguf_metadata.kv_cache_info.cache_dtype})"
@@ -328,9 +222,11 @@ def print_report_for_gguf(
         warnings.warn(
             f"Given that the provided `--model-id {model_id}` (with `--revision {revision}`) is longer than {MAX_DATA_LEN} characters, the table width will be expanded to fit the provided values within their row, but it might lead to unexpected table views. If you'd like to ignore the limit, then provide the `--ignore-table-width` flag to ignore the {MAX_DATA_LEN} width limit, to simply accommodate to whatever the longest text length is."
         )
-    current_len = min(max_centered_len, MAX_DATA_LEN) if ignore_table_width is False else max_centered_len
 
-    # Header
+    current_len = min(max_centered_len, MAX_DATA_LEN) if ignore_table_width is False else max_centered_len
+    max_name_length = min(max(len(filename) for filename in gguf_files.keys()), MAX_DATA_LEN)
+    data_col_width = current_len + 2 * BORDERS_AND_PADDING - max_name_length - 5
+
     _print_header(current_len)
     _print_centered("INFERENCE MEMORY ESTIMATE FOR", current_len)
     _print_centered(f"https://hf.co/{model_id} @ {revision}", current_len)
@@ -339,32 +235,24 @@ def print_report_for_gguf(
             f"w/ max-model-len={first_cache.max_model_len}, batch-size={first_cache.batch_size}",
             current_len,
         )
-    max_name_length = min(max([len(filename) for filename in gguf_files.keys()]), MAX_DATA_LEN)
-    data_col_width = current_len + 2 * BORDERS_AND_PADDING - max_name_length - 5
     _print_divider(data_col_width + 1, "top", name_len=max_name_length)
     _print_row("VERSION", f"hf-mem {__version__}", data_col_width, name_len=max_name_length)
     _print_divider(data_col_width + 1, name_len=max_name_length)
+
     for i, (filename, gguf_metadata) in enumerate(gguf_files.items()):
         transformer = gguf_metadata.components.get("Transformer")
         if transformer and transformer.dtypes:
             if gguf_metadata.kv_cache_info is not None:
-                # Print NAME | GiBs ( PARAMS + KV CACHE) rows
                 gib_text = f"{_bytes_to_gib(transformer.bytes_count + gguf_metadata.kv_cache_info.cache_size):.2f} GiB ({_format_short_number(transformer.param_count)} PARAMS + KV CACHE)"
-                _print_row(
-                    filename + " " * (max_name_length - len(filename)),
-                    gib_text,
-                    data_col_width,
-                    name_len=max_name_length,
-                )
             else:
-                # Print NAME | GiBs ( PARAMS ) rows
                 gib_text = f"{_bytes_to_gib(transformer.bytes_count):.2f} GiB ({_format_short_number(transformer.param_count)} PARAMS)"
-                _print_row(
-                    filename + " " * (max_name_length - len(filename)),
-                    gib_text,
-                    data_col_width,
-                    name_len=max_name_length,
-                )
+
+            _print_row(
+                filename + " " * (max_name_length - len(filename)),
+                gib_text,
+                data_col_width,
+                name_len=max_name_length,
+            )
 
         if i < len(gguf_files) - 1:
             _print_divider(data_col_width + 1, name_len=max_name_length)
